% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{download_raw_data}
\alias{download_raw_data}
\title{Convenience function to download HadUK data from CEDA Archive}
\usage{
download_raw_data()
}
\value{
A tibble containing yearly maxima for each location and each year
}
\description{
This takes aroung 3H30M to run, so don't do it lightly.
The function saves two parquet files in the workding directory that contain
\enumerate{
\item Monthly maxima 2) Yearly maxima
}
}
\details{
The function assumes there are variables CEDA_USR and CEDA_PWD in the R.environ file.
}
