% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.R
\name{fit_gev_trend}
\alias{fit_gev_trend}
\title{A function to calculate and return parameters and Hessian for the generalized extreme distribution with trend}
\usage{
fit_gev_trend(data, priors = "default", ...)
}
\arguments{
\item{data}{The tibble on which to calculate GEV parameters}

\item{y}{The name of the outcome variable in the data}
}
\value{
A tibble with two list columns, one for the parameters and one for the Hessian
}
\description{
A function to calculate and return parameters and Hessian for the generalized extreme distribution with trend
}
