% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{fit_stan}
\alias{fit_stan}
\title{Title}
\usage{
fit_stan(
  data,
  iter_warmup = 500,
  iter_sampling = 500,
  n_chain = 4,
  parallel_chains = 4,
  init = 0,
  refresh = 100
)
}
\arguments{
\item{parallel_chains}{}
}
\description{
Title
}
