% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_priors.R
\name{get_priors}
\alias{get_priors}
\title{Basic data-based priors for theta}
\usage{
get_priors(fit_dat)
}
\arguments{
\item{fit_dat}{Tibble containing a list column with GEV parameters for each station}
}
\value{
A vector containing priors for the latent parameters x
}
\description{
Basic data-based priors for theta
}
