% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{make_download_path}
\alias{make_download_path}
\title{Helper function to insert CEDA FTP username and password into download URLs}
\usage{
make_download_path(filename)
}
\arguments{
\item{filename}{}
}
\description{
Helper function to insert CEDA FTP username and password into download URLs
}
