% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_matrices_spatial.R
\name{make_mu_eta_cond_etahat_spatial}
\alias{make_mu_eta_cond_etahat_spatial}
\title{Title}
\usage{
make_mu_eta_cond_etahat_spatial(Q_eta_cond_etahat)
}
\arguments{
\item{Q_x_cond_etahat}{}
}
\description{
Title
}
