% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_sample-theta.R
\name{met_hast_theta}
\alias{met_hast_theta}
\title{Metropolis Sampler for Theta}
\usage{
met_hast_theta(
  theta,
  Z,
  nu,
  x,
  mu_x,
  chol_Q_x,
  mu_x_cond_etahat,
  chol_Q_x_cond_etahat
)
}
\arguments{
\item{theta}{The current value of theta}

\item{Z}{Model matrix for y as in y = Z * x + e}

\item{x}{The current value of x}

\item{mu_x}{The mean vector for x}

\item{chol_Q_x}{Cholesky decomposition of precision matrix of x}

\item{y}{The current value of y}

\item{chol_Q_e}{Cholesky decomposition of banded diagonal matrix containing the Hessian matrices for each station
from the Max step}

\item{mu_x_cond_y}{Mean vector for x given y}

\item{chol_Q_x_cond_y}{Cholesky decomposition of precision matrix of x given y}
}
\value{
Samples for theta given the previous values of theta
}
\description{
Metropolis Sampler for Theta
}
