% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.R
\name{ms_max}
\alias{ms_max}
\title{Max step of Max-and-Smooth}
\usage{
ms_max(priors = "default")
}
\arguments{
\item{data}{The data on which to perform the Max step}
}
\value{
A tibble with one row per station and nested columns containing results from the Max step
}
\description{
Max step of Max-and-Smooth
}
