% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{ms_smooth_sample}
\alias{ms_smooth_sample}
\title{Helper function for performing sampling}
\usage{
ms_smooth_sample(
  chain,
  p,
  write_to_disk = TRUE,
  type = "spatial",
  theta,
  n_samp = 500
)
}
\arguments{
\item{p}{}
}
\description{
Helper function for performing sampling
}
