% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.R
\name{neg_log_lik_gev_trend}
\alias{neg_log_lik_gev_trend}
\title{Calculate the negative log likelihood of a generalized extreme value distribution with trend using vague priors on the parameters}
\usage{
neg_log_lik_gev_trend(y, t, par, t0 = 1981)
}
\arguments{
\item{y}{The block maxima we want to fit our GEV with trend to}

\item{t}{The time at which each y is measured. To be used with the trend parameter}

\item{par}{The parameters for the GEV with trent}
}
\value{
The negative log likelihood of the outcome, y, at timepoints, t, given the parameters, par.
}
\description{
Calculate the negative log likelihood of a generalized extreme value distribution with trend using vague priors on the parameters
}
