% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.R
\name{p_gev_trend}
\alias{p_gev_trend}
\title{The cumulative distribution function for the generalized extreme value distribution with trend parameter}
\usage{
p_gev_trend(x, t, mu0, sigma, xi, delta, t0 = 1981)
}
\arguments{
\item{x}{The value for which to calculate P(X < x)}

\item{t}{The timepoint at which the value x was measured}

\item{mu0}{The location parameter at timepoint t0}

\item{sigma}{The scale parameter}

\item{xi}{The shape parameter}

\item{delta}{The trend parameter}

\item{t0}{Time time at which mu0 should be applied.}
}
\value{
Returns P(X < x) for the GEV with trend
}
\description{
The cumulative distribution function for the generalized extreme value distribution with trend parameter
}
