% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_distributions_spatial.R
\name{pi_eta_cond_theta_etahat_spatial}
\alias{pi_eta_cond_theta_etahat_spatial}
\title{Log density of x conditioned on etahat and theta}
\usage{
pi_eta_cond_theta_etahat_spatial(
  eta,
  chol_Q_eta_cond_etahat,
  mu_eta_cond_etahat
)
}
\arguments{
\item{x}{The current samples of latent parameters x}
}
\value{
The log density of x when conditioning on y
}
\description{
Log density of x conditioned on etahat and theta
}
