% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_distributions_spatial.R
\name{pi_eta_cond_theta_spatial}
\alias{pi_eta_cond_theta_spatial}
\title{The log density of x conditioned on theta}
\usage{
pi_eta_cond_theta_spatial(eta, chol_Q_e)
}
\arguments{
\item{x}{Current values for the latent parameters x}

\item{mu_x}{Mean vector for x}

\item{chol_Q_x}{Cholesky decomposition of the precision matrix for x}
}
\value{
The log density of x conditioned on theta
}
\description{
The log density of x conditioned on theta
}
