% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_distributions.R
\name{pi_theta}
\alias{pi_theta}
\title{Log prior for Theta}
\usage{
pi_theta(theta)
}
\arguments{
\item{theta}{the current prior parameters for the latent level parameters}
}
\value{
The log density of the parameter vector Theta
}
\description{
Log prior for Theta
}
