% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_distributions_spatial.R
\name{pi_theta_cond_etahat_spatial}
\alias{pi_theta_cond_etahat_spatial}
\title{The log density of Theta conditioned on all other parameters}
\usage{
pi_theta_cond_etahat_spatial(
  theta,
  eta,
  chol_Q_e,
  chol_Q_eta_cond_etahat,
  mu_eta_cond_etahat
)
}
\arguments{
\item{theta}{The vector containing the prior parameters for latent parameters in x}

\item{chol_Q_e}{Cholesky decomposition of block diagonal matrix containing the hessians
for each station from the Max step}

\item{y}{The current samples for estimating the outcome values from the Max step}

\item{x}{The current estimates of the latent level parameters}

\item{mu_x}{The first four parameters in Theta}

\item{chol_Q_x}{Cholesky decomposition of the latter four parameters in Theta}

\item{Z}{Model matrix as in Y = Z * X + e}

\item{mu_x_cond_y}{Same as mu_x except after conditioning on the values of Y}

\item{chol_Q_x_cond_y}{Cholesky decomposition of conditional precision of x after conditioning on y}
}
\value{
The log density of Theta given all other parameters
}
\description{
The log density of Theta conditioned on all other parameters
}
