% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_distributions.R
\name{pi_x_cond_theta}
\alias{pi_x_cond_theta}
\title{The log density of x conditioned on theta}
\usage{
pi_x_cond_theta(x, mu_x, chol_Q_x)
}
\arguments{
\item{x}{Current values for the latent parameters x}

\item{mu_x}{Mean vector for x}

\item{chol_Q_x}{Cholesky decomposition of the precision matrix for x}
}
\value{
The log density of x conditioned on theta
}
\description{
The log density of x conditioned on theta
}
