% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_distributions.R
\name{pi_x_cond_theta_etahat}
\alias{pi_x_cond_theta_etahat}
\title{Log density of x conditioned on etahat and theta}
\usage{
pi_x_cond_theta_etahat(x, mu_x_cond_etahat, chol_Q_x_cond_etahat)
}
\arguments{
\item{x}{The current samples of latent parameters x}
}
\value{
The log density of x when conditioning on y
}
\description{
Log density of x conditioned on etahat and theta
}
