% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_distributions.R
\name{pi_x_cond_theta_y}
\alias{pi_x_cond_theta_y}
\title{Log density of x conditioned on y}
\usage{
pi_x_cond_theta_y(x, mu_x_cond_y, chol_Q_x_cond_y)
}
\arguments{
\item{x}{The current samples of latent parameters x}

\item{mu_x_cond_y}{The mean of x when conditioning on y}

\item{chol_Q_x_cond_y}{Cholesky decomposition of precision matrix of x when conditioning on y}
}
\value{
The log density of x when conditioning on y
}
\description{
Log density of x conditioned on y
}
