% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{process_data}
\alias{process_data}
\title{Helper function to process datafiles downloaded from CEDA Archive}
\usage{
process_data(filename)
}
\arguments{
\item{filename}{}
}
\description{
Helper function to process datafiles downloaded from CEDA Archive
}
