% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_distributions_spatial.R
\name{sample_pi_eta_cond_etahat_spatial}
\alias{sample_pi_eta_cond_etahat_spatial}
\title{Title}
\usage{
sample_pi_eta_cond_etahat_spatial(chol_Q_eta_cond_etahat, mu_eta_cond_etahat)
}
\arguments{
\item{chol_Q_x_cond_etahat}{}
}
\description{
Title
}
