% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_distributions.R
\name{sample_pi_x_cond_etahat}
\alias{sample_pi_x_cond_etahat}
\title{Title}
\usage{
sample_pi_x_cond_etahat(mu_x_cond_etahat, chol_Q_x_cond_etahat)
}
\arguments{
\item{chol_Q_x_cond_etahat}{}
}
\description{
Title
}
