% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_results.R
\name{tidy_helper}
\alias{tidy_helper}
\title{Function to be applied to each chain's results and reduced into one data frame}
\usage{
tidy_helper(datalist)
}
\arguments{
\item{datalist}{}
}
\description{
Function to be applied to each chain's results and reduced into one data frame
}
