% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_results.R
\name{tidy_results}
\alias{tidy_results}
\title{Take the raw results from the MCMC sampler and convert it to a tidy draws_df object}
\usage{
tidy_results(results, type = "spatial")
}
\arguments{
\item{results}{}
}
\description{
Take the raw results from the MCMC sampler and convert it to a tidy draws_df object
}
