% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_params.R
\name{update_theta_basic}
\alias{update_theta_basic}
\title{A function to perform a metropolis hastings step on the proposed theta vector}
\usage{
update_theta_basic(params)
}
\arguments{
\item{theta_prop}{}
}
\description{
A function to perform a metropolis hastings step on the proposed theta vector
}
