% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_params.R
\name{update_theta_spatial}
\alias{update_theta_spatial}
\title{A function to perform a metropolis-hastings step on the proposed theta vector}
\usage{
update_theta_spatial(params)
}
\arguments{
\item{theta_prop}{}
}
\description{
A function to perform a metropolis-hastings step on the proposed theta vector
}
