% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{fit_stan}
\alias{fit_stan}
\title{Use STAN to fit a similar model to the latent gaussian model in Max & Smooth}
\usage{
fit_stan(
  data,
  iter_warmup = 500,
  iter_sampling = 500,
  n_chain = 4,
  parallel_chains = 4,
  init = 0,
  refresh = 100
)
}
\arguments{
\item{data}{The data to use in model fitting}

\item{iter_warmup}{Number of samples in the warmup phase}

\item{iter_sampling}{Number of samples after warmup}

\item{n_chain}{The number of chains}

\item{parallel_chains}{How many chains to sample in parallel}

\item{init}{How to initialize the parameters (See cmdstanr documentation)}

\item{refresh}{How often to print updates (See cmdstanr documentation)}
}
\value{
A cmdstanr model object that has been used to sample from the posterior
}
\description{
Use STAN to fit a similar model to the latent gaussian model in Max & Smooth
}
