% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_priors.R
\name{get_priors}
\alias{get_priors}
\title{Basic data-based priors for theta}
\usage{
get_priors()
}
\arguments{
\item{station_estimates}{Tibble containing a list column with GEV parameters for each station}
}
\value{
A vector containing the prior parameter vector theta
}
\description{
Basic data-based priors for theta
}
