% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{ms_smooth}
\alias{ms_smooth}
\title{Perform the second step, smooth, to sample from the posterior}
\usage{
ms_smooth(
  write_to_disk = TRUE,
  continue_previous = FALSE,
  fmla = station ~ param - 1,
  type = "spatial",
  theta = NULL,
  n_samp = 500,
  n_chain = 4
)
}
\arguments{
\item{fmla}{}

\item{n_samp}{The number of samples total (warmup + sampling).}

\item{n_chain}{The number of chains to sample from.}
}
\value{
Results from the Bayesian samples
}
\description{
Perform the second step, smooth, to sample from the posterior
}
