% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{ms_smooth_sample}
\alias{ms_smooth_sample}
\title{Title}
\usage{
ms_smooth_sample(
  chain,
  p,
  y,
  len_y,
  x,
  len_x,
  Q_e,
  chol_Q_e,
  eta_hat,
  Z,
  priors,
  n_samp = 500
)
}
\arguments{
\item{n_samp}{}
}
\description{
Title
}
